backend/requirements.txt
# Firebase
firebase-admin==6.2.0

# Flask
Flask==3.0.0
Flask-CORS==4.0.0

# Utilities
python-dotenv==1.0.0
APScheduler==3.10.4  # Recurring tasks background job

# Optional: Email
Flask-Mail==0.9.1  # If not using Firebase Cloud Functions
